Edits to Dockerfile to avoid errors in bagdump script
Modify the Dockerfile in that repository to install through apt-get
instead of pip. This means removing packages like NumPy, SciPy, and
Pandas from the command that starts on line 18 and move them to their
apt-get equivalent in the command on line 7. By "apt-get equivalent" I
mean that instead of installing "numpy" install "python-numpy"